queueprocessor/requirements.txt
pymongo
mongo-queue-service
schedule
requests